+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
 
 import { Loading } from ".//pages/Loading";
-import { PrivateApp } from "./components/routes/PrivateApp";
-import { PublicApp } from "./components/routes/PublicApp";
+import { PrivateApp } from "./PrivateApp";
+import { PublicApp } from "./PublicApp";
 import { useUserStore } from "./stores/user";
 
 useUserStore.setState({ fetchingSession: true });
