+++ src/PublicApp.tsx
@@ -2,8 +2,8 @@ import { FC } from "react";
 import { Redirect, Route, Switch } from "react-router-dom";
 import { Container } from "theme-ui";
 
-import { LogIn } from "../../pages/LogIn";
-import { SignUp } from "../../pages/SignUp";
+import { LogIn } from "./pages/LogIn";
+import { SignUp } from "./pages/SignUp";
 
 const PublicAppLayout: FC = ({ children }) => (
   <Container sx={{ maxWidth: 325, pt: 20, textAlign: "center" }}>
